Working?

+++ code/modules/clans/clan_tgui.dm
@@ -1,9 +1,11 @@
 /datum/yautja_panel
-	var/viewed_player = list()
+	var/client/linked_client
 	var/current_menu = "view_clans"
 	var/user_rights = 0
-	var/target_rights = 0
-	var/new_rights = 0
+
+/datum/yautja_panel/New(client/origin_client)
+	. = ..()
+	linked_client = origin_client
 
 /client
 	var/datum/yautja_panel/yautja_panel
@@ -12,9 +14,12 @@
 	set name = "Yautja Clan Panel"
 	set category = "OOC.Records"
 
+	if(!check_whitelist_status(WHITELIST_PREDATOR))
+		return FALSE
+
 	if(yautja_panel)
 		qdel(yautja_panel)
-	yautja_panel = new
+	yautja_panel = new(src)
 	yautja_panel.tgui_interact(mob)
 
 /datum/yautja_panel/tgui_interact(mob/user, datum/tgui/ui)
@@ -56,26 +61,43 @@
 
 	var/datum/entity/clan/formatting_clan
 	var/list/datum/view_record/clan_playerbase_view/clan_view
-
+	var/clan_desc = "This is a list of players without a clan"
 	if(clan_to_format)
 		formatting_clan = GET_CLAN(clan_to_format)
 		formatting_clan.sync()
+		clan_desc = html_encode(formatting_clan.description)
 		clan_view = DB_VIEW(/datum/view_record/clan_playerbase_view, DB_COMP("clan_id", DB_EQUALS, clan_to_format))
 	else
-		clan_view = DB_VIEW(/datum/view_record/clan_playerbase_view, DB_COMP("clan_id", DB_IS, clan_to_format))
+		clan_view = DB_VIEW(/datum/view_record/clan_playerbase_view, DB_COMP("clan_id", DB_IS, null))
 
-	var/list/formatted_clan = list()
+	var/list/members_list = list()
 	for(var/datum/view_record/clan_playerbase_view/CP in clan_view)
+		var/rank_to_give = CP.clan_rank
+		if(CP.permissions & CLAN_PERMISSION_ADMIN_MANAGER)
+			rank_to_give = 999
+
 		var/yautja = list()
 		yautja["ckey"] = CP.ckey
+		yautja["label"] = CP.ckey
+		yautja["name"] = CP.player_name
 		yautja["player_id"] = CP.player_id
 		yautja["rank"] = GLOB.clan_ranks[CP.clan_rank]
 		yautja["honor_amount"] = (CP.honor? CP.honor : 0)
-		//yautja["rank_pos"] = rank_to_give
-		formatted_clan += list(yautja)
+		yautja["rank_pos"] = rank_to_give
+
+		var/datum/entity/player/player = get_player_from_key(CP.ckey)
+		if(player.check_whitelist_status(WHITELIST_YAUTJA_COUNCIL))
+			yautja["label"] = "[CP.ckey] (COUNCILLOR)"
+
+		members_list += list(yautja)
 
 	data["label"] = clan_name
-	data["clan"] = formatted_clan
+	data["desc"] = clan_desc
+	data["color"] = formatting_clan?.color
+	data["honor"] = (formatting_clan?.honor? formatting_clan.honor : 0)
+	data["members"] = members_list
+	if(clan_to_format)
+		data["clan_id"] = clan_to_format
 
 	return data
